requirements-research.txt
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
pyyaml>=6.0
scikit-learn>=1.3
tqdm>=4.66
python-dotenv>=1.0
chars2vec>=0.1.7
transformers>=4.38
torch>=2.1
jupyter>=1.0
notebook>=7.0
pytest>=8.0
